requirements.txt
psycopg2-binary==2.9.9
python-telegram-bot==20.7
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
